+++ extensions/Worker.Extensions.ServiceBus/src/ServiceBusTriggerAttribute.cs
@@ -1,10 +1,15 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-﻿using Microsoft.Azure.Functions.Worker.Extensions.Abstractions;
+using System.Collections.Generic;
+using Azure.Messaging.ServiceBus;
+using Microsoft.Azure.Functions.Worker.Converters;
+using Microsoft.Azure.Functions.Worker.Extensions.Abstractions;
 
 namespace Microsoft.Azure.Functions.Worker
 {
+    [AllowConverterFallback(true)]
+    [InputConverter(typeof(ServiceBusReceivedMessageConverter))]
     public sealed class ServiceBusTriggerAttribute : TriggerBindingAttribute, ISupportCardinality
     {
         private bool _isBatched = false;