+++ extensions/Worker.Extensions.ServiceBus/release_notes.md
@@ -2,3 +2,5 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
+
+- Support binding to ServiceBusReceivedMessage (#1313)